fix: enforce initialize testing with additional assertion
