+++ .gitignore
@@ -10,3 +10,4 @@ examples
 htmlcov
 .pytest_cache
 coverage.xml
+docs
